requirements.txt
selenium
requests
python-dotenv
openai